requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
charset-normalizer==2.0.10
discord.py==1.7.3
idna==2.10
multidict==4.7.6
requests==2.25.1
typing_extensions==4.0.1
urllib3==1.26.2
